requirements.txt
asgiref==3.3.4
astroid==2.5.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
gunicorn==20.1.0
h5py==3.2.1
httpie==2.4.0
idna==2.10
isort==5.8.0
itypes==1.2.0
Jinja2==2.11.3
Keras==2.4.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.20.1
oauthlib==3.1.0
opencv-contrib-python==4.5.1.48
opencv-contrib-python-headless==4.5.1.48
protobuf==3.1.0
psycopg2==2.8.6
pycparser==2.20
Pygments==2.8.1
PyJWT==2.0.1
pylint-django==2.4.3
pylint-plugin-utils==0.6
PySocks==1.7.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
scipy==1.6.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
